STAR-13 General compile fixes: unused imports, TimeUUID instead of UUID, long instead of int gcBefore, replace System calls with Clock.Global, etc. * Add json-simple dependency used in CompactionLogAnalyzer * Add UCS properties to CassandraRelevantProperties * Checkstyle allow certain uses of java.util.concurrent.CompletableFuture in CC code: CompactionManager.BackgroundCompactionCandidate * Restore C* 5.0 methods in CompactionStrategy getPerLevelSizeBytes, isLevelCompaction, and getSSTableCountPerTWCSBucket used in TableStatsHolder * Restore C* 5.0 CompactionStrategy.getEstimatedRemainingTasks used in StreamSession * Restore C* 5.0 BloomFilterTracker lastFalsePositiveCount, lastTruePositiveCount, and lastTrueNegativeCount used in BloomFilterMetrics * Add TableMetrics.bloomFilterFalseRation used in CC ColumnFamilyStore and RealEnvironment

+++ src/java/org/apache/cassandra/db/compaction/PendingRepairManager.java
@@ -317,7 +317,7 @@ class PendingRepairManager
         return ImmutableList.of();
     }
 
-    synchronized Collection<AbstractCompactionTask> getNextBackgroundTasks(int gcBefore)
+    synchronized Collection<AbstractCompactionTask> getNextBackgroundTasks(long gcBefore)
     {
         if (strategies.isEmpty())
             return ImmutableList.of();
