+++ src/java/org/apache/cassandra/db/compaction/TimeWindowCompactionTask.java
@@ -20,6 +20,7 @@ package org.apache.cassandra.db.compaction;
 
 import java.util.Set;
 
+import org.apache.cassandra.db.ColumnFamilyStore;
 import org.apache.cassandra.db.lifecycle.LifecycleTransaction;
 import org.apache.cassandra.io.sstable.format.SSTableReader;
 
@@ -27,7 +28,7 @@ public class TimeWindowCompactionTask extends CompactionTask
 {
     private final boolean ignoreOverlaps;
 
-    public TimeWindowCompactionTask(ColumnFamilyStore cfs, LifecycleTransaction txn, int gcBefore, boolean ignoreOverlaps, TimeWindowCompactionStrategy strategy)
+    public TimeWindowCompactionTask(ColumnFamilyStore cfs, LifecycleTransaction txn, long gcBefore, boolean ignoreOverlaps, TimeWindowCompactionStrategy strategy)
     {
         super(cfs, txn, gcBefore, false, strategy);
         this.ignoreOverlaps = ignoreOverlaps;
